+++ wazo_confd/plugins/user/middleware.py
@@ -137,11 +137,19 @@ class UserMiddleWare:
         user_dict['switchboards'] = switchboards
 
         if agent:
-            agent = self._middleware_handle.get('agent').create(agent, tenant_uuid)
-            self._middleware_handle.get('user_agent_association').associate(
-                user_dict['uuid'], agent['id'], tenant_uuids
+            created_agent = self._middleware_handle.get('agent').create(
+                agent, tenant_uuid
             )
-            user_dict['agent'] = agent
+            self._middleware_handle.get('user_agent_association').associate(
+                user_dict['uuid'], created_agent['id'], tenant_uuids
+            )
+            created_agent['queues'] = agent['queues']
+            for queue in agent['queues']:
+                self._middleware_handle.get('queue_member').associate(
+                    queue, queue['id'], created_agent['id'], tenant_uuids
+                )
+
+            user_dict['agent'] = created_agent
 
         if forwards:
             self._middleware_handle.get('user_forward_association').associate(