+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## IN PROGRESS
 
+* POST on `/users?recursive=true` updated, to provide a way to create/reuse incalls.
+
+## 23.02
+
 * PUT on `/users?recursive=true` added, to provide a way to update fallbacks and forwards for a specific user.
 * POST on `/users?recursive=true` updated, to provide a way to create a user agent associated to existing queues.